C++ was a mistake
Not because the language is terrible in itself.
But because, the language forces you to take everything, and kind of
ditch C, or take nothing (and get nothing).
The whole template subsystem can quickly bite you in the ass if you
try to get too abstract, and the whole 7 default constructors per class
quickly annoyed me to no end.

Anyway, C++ will disappear as quickly as it appeared.

Meanwhile, I was able to execute some arbitrary code on an Android
ARM64 device, so the whole concept can work on mobile devices, which
should be the main user of this work.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>